% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clip}
\alias{clip}
\alias{read_clip_all}
\alias{read_clip_all_wide}
\alias{read_clip_best_lambda}
\title{Read CliP results}
\usage{
read_clip_all(dir)

read_clip_all_wide(dir)

read_clip_best_lambda(dir)
}
\arguments{
\item{dir}{directory in which to seek for the CliP results}
}
\description{
\href{https://github.com/wwylab/CliP}{CliP} is an algorithm for clonal structure
identification through penalizing pairwise differences by Wenyi Wang Lab
at MD Anderson Cancer Center in Houston.
}
\section{Functions}{
\itemize{
\item \code{read_clip_all()}: Read CliP results for all lambdas

\item \code{read_clip_all_wide()}: Read CliP results for all lambdas, in the wider format

\item \code{read_clip_best_lambda()}: Read CliP results for best lambda only

}}
\examples{
dir <- system.file("extdata", "CliP", package = "readthis")
read_clip_best_lambda(dir)

}
