% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{facets}
\alias{facets}
\alias{read_facets_cnvs}
\title{Read FACETS CNV calls}
\usage{
read_facets_cnvs(path, sample_id = path, chrom_convention = "UCSC")
}
\arguments{
\item{path}{Can be either:
\enumerate{
\item path to a single file, sample ID can be passed using sample_id argument
\item vector of file paths, element names will be used as sample IDs
\item directory containing multiple FACETS .csv files, sample IDs will be
guessed from the file names
}}

\item{sample_id}{Sample ID, used with path is a sigle file}

\item{chrom_convention}{UCSC/NCBI/keep}
}
\description{
Reads the variant calls from \href{https://github.com/mskcc/facets/tree/master}{FACETS}
CNV caller
}
\details{
read_facets_csv() reads a single csv file. If sample_id is not provided,
file path is used
}
