% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutect.R
\name{mutect}
\alias{mutect}
\alias{read_mutect_snvs}
\title{Read Mutect variants}
\usage{
read_mutect_snvs(
  path,
  sample_ids = "drop_first",
  PASS_only = TRUE,
  chrom_convention = "UCSC",
  extract_VEP_fields = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Can be either:
\enumerate{
\item path to a single file
\item vector of file paths, element names will be used as patient IDs
\item directory containing multiple Mutect .vcf files, patient IDs will be
guessed from the file names (should follow convention: <patient_ID>.XXX.XXX.vcf)
}}

\item{sample_ids}{Either:
\enumerate{
\item "drop_first"
\item "all"
\item ID(s) of the selected tumor
samples. Default: "drop_first"
}}

\item{PASS_only}{Keep FILTER == PASS variants only?}

\item{chrom_convention}{UCSC/NCBI/keep}

\item{extract_VEP_fields}{If VCF file contains VEP annotations, following
fields will be extracted: Variant_Classification, impact,
gene_symbol and entrez_id (epxerimental, not tested)}

\item{verbose}{Verbose?}
}
\description{
Reads the variant calls from Mutect2 somatic variants caller
}
