% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelka.R
\name{strelka}
\alias{strelka}
\alias{read_strelka_somatic_snvs}
\title{Read Strelka variants}
\usage{
read_strelka_somatic_snvs(
  file,
  sample_id = NULL,
  overwrite_sample_id = NULL,
  PASS_only = TRUE,
  chrom_convention = "UCSC",
  verbose = TRUE
)
}
\arguments{
\item{file}{VCF file to read}

\item{sample_id}{ID(s) of the tumor sample to use. The last ID will be used
if none provided}

\item{overwrite_sample_id}{This argument can be specified to overwrite the
sample_id present in the VCF file}

\item{PASS_only}{Keep FILTER == PASS variants only?}

\item{chrom_convention}{UCSC/NCBI/keep}

\item{verbose}{Verbose?}
}
\description{
Reads the variant calls from Strelka2 Small Variant Caller
}
\examples{
library(readthis)
path <- system.file("extdata", "Strelka", "S1.somatic.snvs.vcf.gz", package = "readthis")
read_strelka_somatic_snvs(
  path,
  sample_id = "TUMOR",
  overwrite_sample_id = "S1",
  verbose = FALSE
)

}
