% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{facets}
\alias{facets}
\alias{read_facets_cnas}
\title{Read FACETS CNA calls}
\usage{
read_facets_cnas(path, sample_id = path, chrom_convention = "UCSC")
}
\arguments{
\item{path}{Can be either:
\enumerate{
\item path to a single file, sample ID can be passed using sample_id argument
\item vector of file paths, element names will be used as sample IDs
\item directory containing multiple FACETS .csv files, sample IDs will be
guessed from the file names
}}

\item{sample_id}{Sample ID, used with path is a single file}

\item{chrom_convention}{UCSC/NCBI/keep}
}
\description{
Reads the variant calls from \href{https://github.com/mskcc/facets/tree/master}{FACETS}
CNA caller
}
\details{
read_facets_csv() reads a single csv file. If sample_id is not provided,
file path is used
}
\examples{
library(readthis)

file1 <- system.file("extdata", "FACETS", "S1.csv", package = "readthis")
read_facets_cnas(file1)

file2 <- system.file("extdata", "FACETS", "S2.csv", package = "readthis")
files <- c(S1 = file1, S2 = file2)
read_facets_cnas(files)

dir <- system.file("extdata", "FACETS", package = "readthis")
read_facets_cnas(dir)

}
