% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.R
\name{read_ascat_files}
\alias{read_ascat_files}
\title{Read ASCAT CNV calls}
\usage{
read_ascat_files(
  path,
  sample_statistics = NULL,
  sample_id = path,
  sample_id_pattern = "(?<=\\\\/)[:alnum:]*(?=\\\\.)",
  chrom_convention = "UCSC"
)
}
\arguments{
\item{path}{Can be either:
\enumerate{
\item path to a single csv file with ASCAT CNV calls
\item tibble with sample_id, cnvs, and (optionally) sample_statistics columns
containing sample_ids and paths.
\item directory containing multiple ASCAT files with "\emph{.csv" and
"}.samplestatistics.*" names.
}

If path is a single csv file, sampleID can be passed in sample_id argument.
If not provided, path will be used in the output sample_id column.
If path is a directory, sample names will be inferred using str_expr() and
sample_id_pattern.}

\item{sample_statistics}{Path to the ".samplestatistics.* file, used when path
leads to a single csv file}

\item{sample_id}{Sample ID, used with path is a sigle file}

\item{sample_id_pattern}{Pattern used to extract sample IDs from the filenames
when path is a directory}

\item{chrom_convention}{UCSC/NCBI/keep}
}
\description{
Reads the CNV variant calls and sample statistics from
\href{https://github.com/VanLoo-lab/ascat}{ASCAT} CNV caller
}
