% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelka.R
\name{strelka}
\alias{strelka}
\alias{read_strelka_somatic_snvs}
\title{Read Strelka variants}
\usage{
read_strelka_somatic_snvs(
  path,
  sample_ids = "drop_first",
  PASS_only = TRUE,
  patient_id_pattern = "(?<=\\\\/)[:alnum:]*(?=\\\\.)",
  chrom_convention = "UCSC",
  verbose = TRUE
)
}
\arguments{
\item{path}{Can be either:
\enumerate{
\item path to a single file
\item vector of file paths, element names will be used as patient IDs
\item directory containing multiple Mutect .vcf files, patient IDs will be
guessed from the file names (should follow convention: <patient_ID>.XXX.XXX.vcf)
}}

\item{sample_ids}{Either:
\enumerate{
\item "drop_first"
\item "all"
\item ID(s) of the selected tumor
samples. Default: "drop_first"
}}

\item{PASS_only}{Keep FILTER == PASS variants only?}

\item{patient_id_pattern}{If path is a dir only: pattern for str_extract()
that should be used to extract the patient_id from the filenames}

\item{chrom_convention}{UCSC/NCBI/keep}

\item{verbose}{Verbose?}
}
\description{
Reads the variant calls from Strelka2 Small Variant Caller
}
\examples{
library(readthis)

file1 <- system.file("extdata", "Strelka", "S1.somatic.snvs.vcf.gz", package = "readthis")
read_strelka_somatic_snvs(file1)

file2 <- system.file("extdata", "Strelka", "S2.somatic.snvs.vcf.gz", package = "readthis")
files <- c(S1 = file1, S2 = file2)
read_strelka_somatic_snvs(files, verbose = FALSE)

dir <- system.file("extdata", "Strelka", package = "readthis")
read_strelka_somatic_snvs(dir)

}
